refactor(scripts): give the UI scripts one shape
They are staying, so they are worth reading. Four things, none of them
touching what any script clicks:

- scripts/tests -> scripts/ui. They are not tests, and the old name kept
  saying they were - to readers and to review tooling alike.
- One browser_page() helper in utils replaces nine different launch
  blocks; some ran slowed down and wide enough for the admin, some did
  not. VITRINA_UI_HEADLESS=1 now turns the window off for the times the
  point is only to fill a database.
- The filer folder name lived in three scripts that have to agree on it,
  and the organization count sat inline; both are named constants now.
- Every script says at the top what it creates and what has to exist
  first - several of them only work after another one has run.

The four scripts that are still one long recorded function stay that
way. Splitting them into steps like the other five is a reasonable next
step, but it cannot be checked without a portal and a browser, so it
belongs to whoever runs them next.
