+++ scripts/ui/README.md
@@ -1,11 +1,14 @@
 # UI scripts
 
-Recorded browser sessions that drive the portal through Chromium. They create content
+Recorded browser sessions that drive the portal through Chromium. They live in `scripts/ui/`
+rather than under a directory called `tests`, which is what they used to be called and what they
+are not. They create content
 (pages, stories, organisations, datasets, users) so the admin can be looked at with real
 data in it.
 
-**These are not tests.** They assert nothing, they run with `headless=False`, and several
-run with `slow_mo` so a person can watch. The automated suite is `tests/`, run with pytest.
+**These are not tests.** They assert nothing, and they run headed and slowed down so a person can
+watch. The automated suite is `tests/`, run with pytest. `VITRINA_UI_HEADLESS=1` turns the browser
+off when the point is only to get content into the database.
 
 They exist because the django-cms admin is iframe- and JavaScript-driven: the rich text
 editor, the versioning toolbar and the publish actions are out of reach for the HTTP client
@@ -25,7 +28,7 @@ playwright install chromium
 Each script then runs on its own:
 
 ```bash
-python scripts/tests/create_organization.py
+python scripts/ui/create_organization.py
 ```
 
 ## What a script expects
