+++ scripts/ui/create_stories.py
@@ -1,8 +1,13 @@
+"""Create and publish stories through the cms toolbar.
+
+Starts from an empty portal; nothing has to exist first.
+"""
+
 import itertools
 import re
 
 from playwright.sync_api import Page, Playwright, sync_playwright
-from utils import BASE_URL, login
+from utils import browser_page
 
 # ---------------------------------------------------------------------------
 # Configuration
@@ -94,17 +99,9 @@ def create_multiple_stories(page: Page, count: int) -> None:
 # ---------------------------------------------------------------------------
 
 def run(playwright: Playwright) -> None:
-    browser = playwright.chromium.launch(headless=False, slow_mo=500)
-    context = browser.new_context(viewport={"width": 1920, "height": 1080})
-    page = context.new_page()
-
-    page.goto(BASE_URL)
-    login(page)
-    go_to_stories_list(page)
-    create_multiple_stories(page, STORY_COUNT)
-
-    context.close()
-    browser.close()
+    with browser_page(playwright) as page:
+        go_to_stories_list(page)
+        create_multiple_stories(page, STORY_COUNT)
 
 
 if __name__ == "__main__":
