+++ scripts/ui/utils.py
@@ -1,6 +1,7 @@
 import os
+from contextlib import contextmanager
 
-from playwright.sync_api import Page
+from playwright.sync_api import Page, Playwright
 
 # ---------------------------------------------------------------------------
 # Configuration
@@ -11,6 +12,16 @@ BASE_URL = os.environ.get("VITRINA_UI_URL", "http://localhost:8000").rstrip("/")
 
 EMAIL_VAR = "VITRINA_UI_EMAIL"
 PASSWORD_VAR = "VITRINA_UI_PASSWORD"
+
+# These scripts are meant to be watched, so the browser is visible and slowed
+# down by default. VITRINA_UI_HEADLESS=1 turns that off when the point is only
+# to get content into the database.
+# The filer folder create_blog_posts.py makes and the other two read from.
+IMAGE_FOLDER = "Skaiciai"
+
+HEADLESS = os.environ.get("VITRINA_UI_HEADLESS") == "1"
+SLOW_MO = 0 if HEADLESS else 500
+VIEWPORT = {"width": 1920, "height": 1080}
 
 
 def _credential(name: str) -> str:
@@ -45,3 +56,23 @@ def login(page: Page, email: str | None = None, password: str | None = None) -> 
     page.get_by_role("textbox", name="Slaptažodis *").fill(password)
     page.get_by_role("button", name="Prisijungti").click()
     page.wait_for_load_state("networkidle")
+
+
+@contextmanager
+def browser_page(playwright: Playwright):
+    """Open a page on the portal, logged in, and close everything afterwards.
+
+    Every script wants the same thing, and before this each one launched the
+    browser its own way - some slowed down, some not, some with a viewport
+    wide enough for the admin and some not.
+    """
+    browser = playwright.chromium.launch(headless=HEADLESS, slow_mo=SLOW_MO)
+    context = browser.new_context(viewport=VIEWPORT)
+    page = context.new_page()
+    page.goto(BASE_URL)
+    login(page)
+    try:
+        yield page
+    finally:
+        context.close()
+        browser.close()
